updated stock property from integer to unsigned big integer

+++ database/migrations/2024_10_07_085253_add_stock_to_products_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('products', function (Blueprint $table) {
             // adding teh stock column to the products table
-            $table->integer('stock')->default(0);
+            $table->unsignedBigInteger('stock')->default(0);
         });
     }
